# libs/shared/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Booking Platform Microservices API
  version: 1.0.0
  description: Complete API specification for Payment, Reservation, Ledger, and Notification services
servers:
  - url: http://localhost:3001
    description: Payment Service
  - url: http://localhost:3002
    description: Reservation Service
  - url: http://localhost:3003
    description: Ledger Service
  - url: http://localhost:3004
    description: Notification Service
paths:
...
  /api/v1/payments/initiate:
    post:
      tags: [Payment]
      summary: Initiate a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitiatePaymentRequest'
      responses:
        '200':
          description: Payment initiated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiatePaymentResponse'
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /api/v1/payments/{paymentId}/confirm:
    post:
      tags: [Payment]
      summary: Confirm a payment
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmPaymentRequest'
      responses:
        '200':
          description: Payment confirmed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmPaymentResponse'
  /api/v1/payments/{paymentId}:
    get:
      tags: [Payment]
      summary: Get payment status
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentStatus'
...
components:
  schemas:
...
    InitiatePaymentRequest:
      type: object
      required: [bookingId, amount, currency, provider, customerPhone]
      properties:
        bookingId:
          type: string
        amount:
          type: number
        currency:
          type: string
          enum: [BDT, USD]
        provider:
          $ref: '#/components/schemas/PaymentProvider'
        customerPhone:
          type: string
        customerEmail:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
    InitiatePaymentResponse:
      type: object
      properties:
        paymentId:
          type: string
        transactionId:
          type: string
        status:
          $ref: '#/components/schemas/PaymentStatus'
        redirectUrl:
          type: string
        errorMessage:
          type: string
    ConfirmPaymentRequest:
      type: object
      required: [transactionId, providerReference]
      properties:
        transactionId:
          type: string
        providerReference:
          type: string
        providerData:
          type: object
          additionalProperties:
            type: string
    ConfirmPaymentResponse:
      type: object
      properties:
        paymentId:
          type: string
        status:
          $ref: '#/components/schemas/PaymentStatus'
        amount:
          type: number
        errorMessage:
          type: string
    PaymentStatus:
      type: string
      enum: [PENDING, PROCESSING, COMPLETED, FAILED, CANCELLED, REFUNDED]
    PaymentProvider:
      type: string
      enum: [BKASH, NAGAD, SSLCOMMERZ]

    # Reservation Schemas
...